Add files via upload
These Arduino files are the start of our Arduino button mapping. KeyboardTest.ino is an example I found online that shows how to make an Arduino UNO or MEGA act as a keyboard. shifter.ino is the start of the Arduino code that will implement electronic shifting and readFromAnalog.ino is a simple program to print out the value read from analog input, so that we might be able to determine what signals the shifter puts out.